+++ src/data/pt/certificationsData.tsx
@@ -1,7 +1,6 @@
-// src/data/pt/certificationsData.ts
 import React from "react";
 import Highlight from "@/components/Highlight";
-import { CertificationItem } from "../types"; // UPDATED IMPORT
+import { CertificationItem } from "../en/certificationsData";
 
 export const certificationsData: CertificationItem[] = [
   {
@@ -96,7 +95,7 @@ export const certificationsData: CertificationItem[] = [
     id: "tryhackme-learning-paths",
     name: "Percursos de Aprendizagem no TryHackMe",
     issuer: "TryHackMe",
-    credentialUrl: "https://tryhackme.com/p/joao.grilo.dev",
+    credentialUrl: "https://tryhackme.com/p/joaoptgrilo",
     description: (
       <>
         A progredir em percursos de cibersegurança fundamentais, incluindo{" "}
