+++ src/data/pt/experienceData.tsx
@@ -1,7 +1,6 @@
-// src/data/pt/experienceData.ts
 import React from "react";
 import Highlight from "@/components/Highlight";
-import { ExperienceItem } from "../types"; // UPDATED IMPORT
+import { ExperienceItem } from "../en/experienceData";
 
 export const experienceData: ExperienceItem[] = [
   {
@@ -46,6 +45,7 @@ export const experienceData: ExperienceItem[] = [
         reprodutíveis.
       </>,
     ],
+    // THIS IS THE CORRECTED SECTION WITH DESCRIPTIONS FOR TOOLTIPS
     keyTech: [
       {
         name: "PHP",
@@ -117,6 +117,7 @@ export const experienceData: ExperienceItem[] = [
         <Highlight>reduções médias de ~40% no tempo de carregamento</Highlight>.
       </>,
     ],
+    // THIS IS THE CORRECTED SECTION WITH DESCRIPTIONS FOR TOOLTIPS
     keyTech: [
       {
         name: "PHP",
